feat(wakatime): implemented WakaTime validation for retrieving WakaTime Bar/Pie Data
I noticed a bug where if a user pasted an URL for the wrong input bar (i.e. Languages' URL to
TimeBar Input). If the mistake happened, the response from the server for both the WakaTimeBarGraph
and WakaTimePieGraph were not appropriate. So to fix this problem, I added validation for both
instances to make sure that the data being retrieved from the URL is valid for either graphs.